+++ .gitignore
@@ -14,6 +14,7 @@ build
 .DS_STORE
 tests/fixtures/.acquia/acquia-cli.json
 tests/fixtures/.acquia/cloud_api.conf
+tests/fixtures/project/.acquia-cli.yml
 
 ###> squizlabs/php_codesniffer ###
 /.phpcs-cache
